Moved the enable-system-look-and-feel function to AbstractSwingView
